// src/app/core/services/models/update-profile-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UpdateProfileRequest {
  dateOfBirth?: string;
  email: string;
  firstname: string;
  gender?: 'MALE' | 'FEMALE';
  lastname: string;
  skill?: 'COMMUNICATION' | 'ORGANIZATION' | 'IT' | 'TEACHING' | 'FIRST_AID' | 'LOGISTICS' | 'PSYCHOLOGICAL_SUPPORT' | 'LEADERSHIP';
}